Moved extract_units function into parsing and updated tests.
